default kwargs set to None
You want to avoid having the same dict used for every future function call. The dict in the kwargs will be evaluated on function definition, not function call. So if you do some_kwargs={} you'll end up with the same {} used every time the function is called. Better to avoid it.