get_prep_value returns default instead of None
Changed get_prep_value to return get_default() instead of None. This was preventing the field from being non-nullable, as get_prep_value would always return None during the save process, preventing the field from using a default of an empty string (""), as both an empty string and None would evaluate to false in the `if not value` condition.