Handle "not found" error while renaming
The call_api function of the uploader module does not raise specific
error, but converts all exceptions into the Error class. This approach
makes it very difficult to handle errors appropriately. The solution
to this problem might be to raise dedicated errors (same as the API
module). However, for now, the simplest solution is to add the 404 code
into the list of known codes.